requirements.txt
icalendar
requests
http://cdn.mysql.com/Downloads/Connector-Python/mysql-connector-python-2.0.4.zip#md5=3df394d89300db95163f17c843ef49df
sqlalchemy
schedule
python-telegram-bot